refactor(core): move getCard to helpers to break effects/validActions circular dependency
- Add engine/helpers/cardLookup.ts with getCard (pure lookup, data-layer imports)
- Update effects: resolvability, describeEffect, cardBoostEffects, cardBoostResolvers, effectUndoContext to import from ../helpers/cardLookup.js
- Remove getCard impl from validActions/cards/index.ts; re-export from helpers for backward compatibility
- Export getCard from engine/helpers/index.ts

effects/ no longer imports from validActions/; both use helpers/cardLookup.
